Add google oauth login (#1)
This will go through google oauth process, receive an id token and send
to the API.  The API will decode and either find or create a user in the
database with the google user's email and return a JWT that is stored in
local storage. The JWT should be used as a bearer token for all requestes.

What's great about google oauth is we can grab people's profile pic.
:+1:

TODO: should refactor web request stuff to always use bearer token from
local storage.