0.1.0 Production Deployment (#38)
* Use html_url from GH api

* Make some ambiguous improvements

* Set configuration variables for oauth

* Set api url in staging environment

* Fix authentication woes

The common.js file was being loaded and setting the default headers to
use a Bearer token of null for all future requests (if a token did not
exist in localStorage).

* Check authentication based off of picture

There is actually a moment in time where the JWT in localStorage is from
Google.  Thus, a user is not really authenticated at that point.